+++ atomic_heart/atomic_heart.py
@@ -1,32 +1,33 @@
 """Atomic Heart test script"""
 
-from argparse import ArgumentParser
 import logging
 import os
-from pathlib import Path
-import time
 import sys
+import time
+from argparse import ArgumentParser
+from pathlib import Path
+
 import pydirectinput as user
-from utils import read_resolution
+from atomic_heart_utils import read_resolution
 
 PARENT_DIRECTORY = str(Path(__file__).resolve().parent.parent)
 sys.path.insert(1, PARENT_DIRECTORY)
 
+from harness_utils.artifacts import ArtifactManager, ArtifactType
+from harness_utils.keras_service import KerasService
+from harness_utils.misc import press_n_times, remove_files
+from harness_utils.output import (
+    format_resolution,
+    seconds_to_milliseconds,
+    setup_logging,
+    write_report_json,
+)
+from harness_utils.process import terminate_processes
 from harness_utils.steam import (
     exec_steam_run_command,
     get_app_install_location,
     get_build_id,
 )
-from harness_utils.misc import remove_files, press_n_times
-from harness_utils.process import terminate_processes
-from harness_utils.output import (
-    setup_logging,
-    format_resolution,
-    seconds_to_milliseconds,
-    write_report_json,
-)
-from harness_utils.keras_service import KerasService
-from harness_utils.artifacts import ArtifactManager, ArtifactType
 
 SCRIPT_DIRECTORY = Path(__file__).resolve().parent
 LOG_DIRECTORY = SCRIPT_DIRECTORY / "run"
